cleaning up the code a bit and adding some documentation for historical choices
